# Removed auto-decoding from the single needle version of startsWith