@CyberShadow 2017-05-24
> Removed auto-decoding from the single argument version of startsWith

This PR does not modify the single-argument overload. The single-argument overload (`bool startsWith(alias pred, R)(R doesThisStart)`) is below the overload modified here.