@nordlow 2017-05-24
Can we add the same optimization to `std.algorithm.searching.find`? :)